corrBarplot

This function creates a either a static or interactive bar plot of correlations between variables in a dataset.

corrBarplot(
  data,
  method = c("pearson", "kendall", "spearman"),
  interactive = TRUE,
  pal = colorRampPalette(c("cornflowerblue", "white", "tomato"))(100)
)

Arguments

data

A dataframe containing the data to be analyzed.

method

A character string specifying the correlation method. One of "pearson", "kendall", or "spearman". Default is "pearson".

interactive

A logical value specifying whether to create an interactive ggplotly plot, default is TRUE

pal

A colour palette for the bar plot, default is colorRampPalette(c("cornflowerblue", "white", "tomato"))(100).

Value

A static or interactive bar plot displaying correlations.

Details

Creates a static or interactive bar plot displaying correlation values. By hovering mouse over a bar, the variables and correlation value is shown.

Examples

corrBarplot(data = mtcars,
            method = 'pearson',
            interactive = TRUE)