corrBubble

This function creates an interactive bubble plot of correlations between variables in a dataset.

corrBubble(
  data,
  method = c("pearson", "kendall", "spearman"),
  display = c("all", "upper", "lower"),
  pal = colorRampPalette(c("cornflowerblue", "white", "tomato"))(100)
)

Arguments

data

A dataframe containing the data to be analyzed.

method

A character string specifying the correlation method. One of "pearson", "kendall", or "spearman". Default is "pearson".

display

A character vector specifying which part of the correlation matrix to display: 'all', 'upper', or 'lower', default is 'all'.

pal

A color palette for the bubble plot.

Value

An interactive bubble plot displaying correlations.

Details

Creates an interactive bubble plot displaying correlation values. By hovering mouse over a cell, the variables and correlation value is shown.

Examples

corrBubble(data = mtcars,
           method = 'pearson',
           display = 'all')