corrHeatmap

Creates an interactive heatmap plot displaying correlations.

corrHeatmap(
  data,
  method = c("pearson", "kendall", "spearman"),
  display = c("all", "upper", "lower"),
  reorder = TRUE,
  pal = colorspace::diverging_hcl(palette = "Blue-Red", n = 100)
)

Arguments

data

A data frame.

method

Which correlation coefficient (or covariance) is to be computed. One of "pearson" (default), "kendall", or "spearman": can be abbreviated.

display

Choose to display the upper, lower, or full matrix of values using "upper", "lower", or "all", respectively.

reorder

If TRUE (the default) then the heatmap is reordered to group correlations.

pal

The colour palette to use for displaying values.

Value

An interactive heatmap plot displaying correlations.

Details

Creates an interactive heatmap displaying correlation values. By hovering mouse over a cell, the variables and correlation value is shown.

Examples

corrHeatmap(data = mtcars, display = 'all')