corrPairs

This function creates a pairwise correlation plot with annotated correlation coefficients and optional coloring by a specified variable. The plot can be interactive or static.

corrPairs(
  data,
  method = c("pearson", "kendall", "spearman"),
  interactive = TRUE,
  col_by = NULL
)

Arguments

data

A data frame containing the variables to be plotted.

method

A character string specifying the correlation method. One of "pearson", "kendall", or "spearman". Default is "pearson".

interactive

A logical value indicating whether the output plot should be interactive (TRUE) or static (FALSE). Default is TRUE.

col_by

An optional character string specifying the name of the column in the data frame to be used for coloring points. Default is NULL.

Value

A ggplotly object (if interactive = TRUE) or a ggplot object (if interactive = FALSE) displaying the pairwise correlation plot with annotated correlation coefficients and optional coloring by the specified variable.

Examples

corrPairs(data = mtcars[,1:4],
         method = "pearson",
         interactive = TRUE,
         col_by = "cyl")
#> Warning: The following aesthetics were dropped during statistical transformation:
#> x_plotlyDomain
#> ℹ This can happen when ggplot fails to infer the correct grouping structure in
#>   the data.
#> ℹ Did you forget to specify a `group` aesthetic or to convert a numerical
#>   variable into a factor?
#> Warning: Ignoring unknown aesthetics: text
#> Warning: The following aesthetics were dropped during statistical transformation:
#> x_plotlyDomain, y_plotlyDomain
#> ℹ This can happen when ggplot fails to infer the correct grouping structure in
#>   the data.
#> ℹ Did you forget to specify a `group` aesthetic or to convert a numerical
#>   variable into a factor?
#> Warning: Ignoring unknown aesthetics: text
#> Warning: The following aesthetics were dropped during statistical transformation:
#> x_plotlyDomain, y_plotlyDomain
#> ℹ This can happen when ggplot fails to infer the correct grouping structure in
#>   the data.
#> ℹ Did you forget to specify a `group` aesthetic or to convert a numerical
#>   variable into a factor?
#> Warning: Ignoring unknown aesthetics: text
#> Warning: The following aesthetics were dropped during statistical transformation:
#> x_plotlyDomain, y_plotlyDomain
#> ℹ This can happen when ggplot fails to infer the correct grouping structure in
#>   the data.
#> ℹ Did you forget to specify a `group` aesthetic or to convert a numerical
#>   variable into a factor?
#> Warning: The following aesthetics were dropped during statistical transformation:
#> x_plotlyDomain
#> ℹ This can happen when ggplot fails to infer the correct grouping structure in
#>   the data.
#> ℹ Did you forget to specify a `group` aesthetic or to convert a numerical
#>   variable into a factor?
#> Warning: Ignoring unknown aesthetics: text
#> Warning: pseudoinverse used at 3.98
#> Warning: neighborhood radius 4.02
#> Warning: reciprocal condition number  2.0055e-16
#> Warning: There are other near singularities as well. 16.16
#> Warning: The following aesthetics were dropped during statistical transformation:
#> x_plotlyDomain, y_plotlyDomain
#> ℹ This can happen when ggplot fails to infer the correct grouping structure in
#>   the data.
#> ℹ Did you forget to specify a `group` aesthetic or to convert a numerical
#>   variable into a factor?
#> Warning: Ignoring unknown aesthetics: text
#> Warning: pseudoinverse used at 3.98
#> Warning: neighborhood radius 4.02
#> Warning: reciprocal condition number  2.0055e-16
#> Warning: There are other near singularities as well. 16.16
#> Warning: The following aesthetics were dropped during statistical transformation:
#> x_plotlyDomain, y_plotlyDomain
#> ℹ This can happen when ggplot fails to infer the correct grouping structure in
#>   the data.
#> ℹ Did you forget to specify a `group` aesthetic or to convert a numerical
#>   variable into a factor?
#> Warning: Can only have one: highlight
#> Warning: The following aesthetics were dropped during statistical transformation:
#> x_plotlyDomain
#> ℹ This can happen when ggplot fails to infer the correct grouping structure in
#>   the data.
#> ℹ Did you forget to specify a `group` aesthetic or to convert a numerical
#>   variable into a factor?
#> Warning: Ignoring unknown aesthetics: text
#> Warning: The following aesthetics were dropped during statistical transformation:
#> x_plotlyDomain, y_plotlyDomain
#> ℹ This can happen when ggplot fails to infer the correct grouping structure in
#>   the data.
#> ℹ Did you forget to specify a `group` aesthetic or to convert a numerical
#>   variable into a factor?
#> Warning: Can only have one: highlight
#> Warning: The following aesthetics were dropped during statistical transformation:
#> x_plotlyDomain
#> ℹ This can happen when ggplot fails to infer the correct grouping structure in
#>   the data.
#> ℹ Did you forget to specify a `group` aesthetic or to convert a numerical
#>   variable into a factor?
#> Warning: Can only have one: highlight