corrShiny

Correlation Explorer Shiny App. This function creates a Shiny application to explore the correlation between variables in a given dataset.

corrShiny(
  data,
  x_var,
  y_var,
  color_var = NULL,
  size_var = NULL,
  correlation_method = "pearson"
)

Arguments

data

A data frame with the variables to be analyzed.

x_var

The name of the variable to be plotted on the X-axis.

y_var

The name of the variable to be plotted on the Y-axis.

color_var

The name of the variable to be used for coloring the points on the scatter plot.

size_var

The name of the variable to be used for sizing the points on the scatter plot.

correlation_method

The method to be used for computing the correlation coefficient, must be one of "pearson", "spearman" or "kendall".

Value

A Shiny app that displays a scatter plot and the correlation coefficient between two variables.