corrBubble

This function creates an interactive bubble plot of correlations between variables in a dataset.

corrBubble(
  mat,
  display = c("all", "upper", "lower"),
  pal = colorRampPalette(c("cornflowerblue", "white", "tomato"))(100)
)

Arguments

mat

A square correlation matrix to visualise.

display

A character vector specifying which part of the correlation matrix to display: 'all', 'upper', or 'lower', default is 'all'.

pal

A color palette for the bubble plot.

Value

An interactive bubble plot displaying correlations.

Details

Creates an interactive bubble plot displaying correlation values. By hovering mouse over a cell, the variables and correlation value is shown.

Examples

cm <- cor(mtcars)

corrBubble(mat = cm,
           display = 'all')