corrHeatmap

This function creates an interactive heatmap of correlations between variables in a dataset.

corrHeatmap(
  data,
  method = c("pearson", "kendall", "spearman"),
  display = c("all", "upper", "lower"),
  reorder = TRUE,
  pal = colorRampPalette(c("darkblue", "white", "darkred"))(100)
)

Arguments

data

A dataframe containing the data to be analyzed.

method

A character string specifying the correlation method. One of "pearson", "kendall", or "spearman". Default is "pearson".

display

A character vector specifying which part of the correlation matrix to display: 'all', 'upper', or 'lower', default is 'all'.

reorder

A logical value indicating whether to reorder the heatmap based on hierarchical clustering, default is TRUE.

pal

A color palette for the heatmap.

Value

An interactive heatmap plot displaying correlations.

Details

Creates an interactive heatmap displaying correlation values. By hovering mouse over a cell, the variables and correlation value is shown.

Examples

corrHeatmap(data = mtcars,
             method = 'pearson',
             display = 'all')